+++ lib/mechanics/supervisor.ml
@@ -90,11 +90,11 @@ end = struct
       in
       let rec go ~depth = function
         | Call (f, _args) as t when is_extractable f && depth > 0 -> extract t
-        | Call (op, args) when not (Symbol.is_lazy op) ->
+        | Call (op, args) when not (Symbol.is_lazy_op op) ->
           go_args ~depth ~op ~acc:Fun.id args
         | Var _ | Const _ | Call _ -> ()
       and go_args ~depth ~op ~acc = function
-        | [] -> go_call ~depth ~op (Symbol.kind op, acc [])
+        | [] -> go_call ~depth ~op (Symbol.op_kind op, acc [])
         | t :: rest when is_value t ->
           go_args ~depth ~op ~acc:(fun xs -> acc (t :: xs)) rest
         | t :: rest ->
@@ -103,7 +103,7 @@ end = struct
              let shell = Call (op, acc [] @ [ shell ] @ rest) in
              raise_notrace (Extract ((x, call), shell)))
       and go_call ~depth ~op = function
-        | `GCall, Call (c, _c_args) :: _args when Symbol.kind c = `CCall -> ()
+        | `GCall, Call (c, _c_args) :: _args when Symbol.op_kind c = `CCall -> ()
         | `GCall, (_ :: _ as args) when is_unproductive op && depth > 0 ->
           extract (Call (op, args))
         | (`CCall | `FCall | `GCall), _args -> ()